nm: If the file is ET_REL and we need Dwarf use libdwfl to get it relocated.
If we need Dwarf and the file is ET_REL use the same trick as in readelf
to get a relocated Dwarf. Otherwise lots of references in the debug_info
will come out as zero.

This also explains the "bogus" Dwarf seen that caused the memory leak in
the local_root that was fixed previously.

Signed-off-by: Mark Wielaard <mjw@redhat.com>