config: Fix check for argp_parse to pass &argv
Right now it's passing a char* when it expects a char** instead.

This usually produces a warning that may go unnoticed, but if CFLAGS
contains -Werror, that breaks the ./configure run with the following
error:

  $ ./configure CFLAGS=-Werror
  ...
  configure: WARNING: "libc does not have argp"
  checking for argp_parse in -largp... no
  configure: error: "no libargp found"

Tested: Checked that after this fix, running ./configure CFLAGS=-Werror
works as expected and argp_parse is correctly detected.

Signed-off-by: Filipe Brandenburger <filbranden@google.com>